Linux CI

+++ build/unix.sh
@@ -10,8 +10,8 @@ if [ "$(uname)" == "Darwin" ]; then
     EXIV2_LIB_DIST_NAME="libexiv2.27.dylib"
 else
     EXIV2_BUILD_NAME="exiv2-0.27.2-Linux64"
-    EXIV2_LIB_NAME="libexiv2.so.0.27.2"
-    EXIV2_LIB_DIST_NAME="libexiv2.so.27"
+    EXIV2_LIB_NAME="-"
+    EXIV2_LIB_DIST_NAME="-"
 fi
 
 
@@ -21,7 +21,8 @@ mkdir -p exiv2
 tar xzf exiv2.tar.gz -C exiv2
 
 mv "exiv2/$EXIV2_BUILD_NAME/bin/exiv2json" exiv2/exiv2json
-mv "exiv2/$EXIV2_BUILD_NAME/lib/$EXIV2_LIB_NAME" "exiv2/$EXIV2_LIB_DIST_NAME"
-ls -al exiv2
+if [ -e "exiv2/$EXIV2_BUILD_NAME/lib/$EXIV2_LIB_NAME" ]; then
+    mv "exiv2/$EXIV2_BUILD_NAME/lib/$EXIV2_LIB_NAME" "exiv2/$EXIV2_LIB_DIST_NAME"
+fi
 
 rm -rf "exiv2/$EXIV2_BUILD_NAME" exiv2.tar.gz
